feat(feed): extract reorderFeedItems utility with rule ④ content filter
