+++ apps/product/src/app/[locale]/(with-layout)/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { type ActivityCardItem, type FeedItem, useFeed, useReactionsBatch } from "@daodao/api";
+import { type FeedItem, reorderFeedItems, useFeed, useReactionsBatch } from "@daodao/api";
 import { useAuth } from "@daodao/auth";
 import { useTranslations } from "@daodao/i18n";
 import { useRouter, useSearchParams } from "@daodao/i18n/navigation";
@@ -20,75 +20,6 @@ import {
   ShowcaseSearchBar,
 } from "@/components/showcase";
 import { HOME_TAB_PATHS } from "@/constants/home-navigation";
-
-// Reorder feed items into the cycle: [打卡] → [互動] → [實踐] → repeat (1:1:1)
-// 規則：
-// ① 最新打卡和互動排前面（API 已依時間排序，保留 bucket 內相對順序）
-// ② 自己的互動不優先（currentUserId 的互動推後）
-// ③ 有打卡的實踐不再顯示實踐卡（同一 practice_id 只擇一）
-// ⑤ 打卡:互動:實踐 = 1:1:1
-function reorderFeedItems(items: FeedItem[], currentUserId?: string | null): FeedItem[] {
-  // Rule ③: 蒐集有打卡的 practice_id
-  const practiceIdsWithCheckins = new Set<string>();
-  for (const item of items) {
-    if (item.type === "checkin") {
-      practiceIdsWithCheckins.add(item.data.practice.id);
-    }
-  }
-
-  const checkins: Extract<FeedItem, { type: "checkin" }>[] = [];
-  const interactions: FeedItem[] = [];
-  const practices: Extract<FeedItem, { type: "practice" }>[] = [];
-
-  for (const item of items) {
-    if (item.type === "checkin") {
-      if (item.feed_reason === "cheered") {
-        interactions.push(item);
-      } else {
-        checkins.push(item);
-      }
-    } else if (item.type === "practice") {
-      if (item.feed_reason === "cheered") {
-        interactions.push(item);
-      } else if (!practiceIdsWithCheckins.has(item.data.id)) {
-        // Rule ③: 有打卡紀錄的實踐不出現實踐卡
-        practices.push(item);
-      }
-    } else if (item.type === "activity") {
-      interactions.push(item as ActivityCardItem);
-    }
-  }
-
-  // Rule ②: 自己的互動推到後面
-  if (currentUserId) {
-    const isOwn = (item: FeedItem) =>
-      (item.type === "checkin" || item.type === "practice") && item.data.user?.id === currentUserId;
-    const others = interactions.filter((i) => !isOwn(i));
-    const own = interactions.filter((i) => isOwn(i));
-    interactions.length = 0;
-    interactions.push(...others, ...own);
-  }
-
-  // Rule ⑤: 1:1:1 cycle
-  const result: FeedItem[] = [];
-  let ci = 0;
-  let ii = 0;
-  let pi = 0;
-
-  while (ci < checkins.length || ii < interactions.length || pi < practices.length) {
-    if (ci < checkins.length) {
-      result.push(checkins[ci++]!);
-    }
-    if (ii < interactions.length) {
-      result.push(interactions[ii++]!);
-    }
-    if (pi < practices.length) {
-      result.push(practices[pi++]!);
-    }
-  }
-
-  return result;
-}
 
 export default function HomePage() {
   const router = useRouter();
